# Refactor streaming watch encoder to enable caching