Fix: https://github.com/kubernetes/kubernetes/issues/110146

```release-note
NONE
```

I used the benchmark that was originally added in https://github.com/kubernetes/kubernetes/pull/113326 to benchmark this change.  It seems it's achieving what we wanted:
```
wojtekt@wojtekt-l:~/go/src/k8s.io/kubernetes$ benchcmp old.txt new.txt 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                 old ns/op     new ns/op     delta
BenchmarkWatchHTTP-8                      36086         34110         -5.48%
BenchmarkWatchHTTP_UTF8-8                 34019         33528         -1.44%
BenchmarkWatchWebsocket-8                 37032         37368         +0.91%
BenchmarkWatchProtobuf-8                  14550         15401         +5.85%
BenchmarkWatchCachingObjectJSON-8         23124         10613         -54.10%
BenchmarkWatchCachingObjectProtobuf-8     12834         10562         -17.70%
```

For completeness also memory [cpu should be ignored here]
```
before:
BenchmarkWatchHTTP-8                    	   33645	     32539 ns/op	     835 B/op	      12 allocs/op
BenchmarkWatchHTTP_UTF8-8               	   32142	     34380 ns/op	     835 B/op	      12 allocs/op
BenchmarkWatchWebsocket-8               	   33790	     35212 ns/op	    3186 B/op	      20 allocs/op
BenchmarkWatchProtobuf-8                	   73652	     15181 ns/op	     736 B/op	       7 allocs/op
BenchmarkWatchCachingObjectJSON-8       	   51310	     25799 ns/op	     152 B/op	       6 allocs/op
BenchmarkWatchCachingObjectProtobuf-8   	   89762	     12142 ns/op	     120 B/op	       4 allocs/op
PASS

after:
BenchmarkWatchHTTP-8                    	   36710	     34183 ns/op	     835 B/op	      12 allocs/op
BenchmarkWatchHTTP_UTF8-8               	   35911	     29547 ns/op	     835 B/op	      12 allocs/op
BenchmarkWatchWebsocket-8               	   32452	     37185 ns/op	    3187 B/op	      20 allocs/op
BenchmarkWatchProtobuf-8                	   81285	     15854 ns/op	     736 B/op	       7 allocs/op
BenchmarkWatchCachingObjectJSON-8       	   46171	     26953 ns/op	     152 B/op	       6 allocs/op
BenchmarkWatchCachingObjectProtobuf-8   	   94173	     13723 ns/op	     120 B/op	       4 allocs/op
PASS
```


The first 3 tests are fluctuating across runs, but the CachingObject ones are consistently better.

/kind cleanup
/priority important-longterm
/sig api-machinery