@wojtek-t 2023-09-11
The problems with scale tests were fixed on Friday

/test pull-kubernetes-e2e-gce-scale-performance-manual