@wojtek-t 2023-09-11
Should be fixed with https://github.com/kubernetes/test-infra/pull/30674

/test pull-kubernetes-e2e-gce-scale-performance-manual