@wojtek-t 2023-09-15
/assign @liggitt 

Would you be able to help with reviewing this as a follow up from https://github.com/kubernetes/kubernetes/pull/119801 ?