@p0lyn0mial 2023-09-19
> The numbers look worse if we don't use CachingObject, but given we use that for all events except from initEvents, the numbers there are better.

FYI: I was planning to enable CachingObject for initEvens once we have streaming