@wojtek-t 2023-09-25
> FYI: I was planning to enable CachingObject for initEvens once we have streaming

We should chat about it - it's not obvious if it will cause more harm the gain.