@liggitt 2023-10-24
Do the benchmarks capture the increased retained memory cost of this change? If I understand correctly, each object emitted in a watch event would now have at least two serialized copies cached (its individual serialization, then its embedded-in-watch-event serialization), is that right?

That reduces alloc / cpu churn emitting the same object to multiple watchers at the cost of ~doubling (?) the steady state memory cost of the watch cache, right? Are there usage patterns where this tradeoff isn't worth it (e.g. single/small numbers of watchers), or would even be prohibitive (things currently running very close to memory limits)? Do we need a way to opt out of this change?