@wojtek-t 2023-10-24
> Do the benchmarks capture the increased retained memory cost of this change? If I understand correctly, each object emitted in a watch event would now have at least two serialized copies cached (its individual serialization, then its embedded-in-watch-event serialization), is that right?

@liggitt - that's not fully true, we're not retaining that (in objects that are stored in watch cache)
the serializations are only kept in memory for the time of emitting a given object to watchers registered for it in a given point in time. it's immediately dropped after it's emitted.
So we're generally talking about <1s of having that stored in memory.

So it's definitely not doubling the memory - we're rather talking about minor throughput-proportional increase.
So I don't see a concern even for clusters running close to memory limits.