@liggitt 2023-10-26
> @liggitt - that's not fully true, we're not retaining that (in objects that are stored in watch cache)
> the serializations are only kept in memory for the time of emitting a given object to watchers registered for it in a given point in time. it's immediately dropped after it's emitted.

Ah, ok... that wasn't at all clear to me. The object we emit to watchers from the watch cache doesn't come through here? If it *did*, and this *did* start retaining / doubling memory long-term, would anything catch that?