@wojtek-t 2023-10-27
> Ah, ok... that wasn't at all clear to me. The object we emit to watchers from the watch cache doesn't come through here? If it did, and this did start retaining / doubling memory long-term, would anything catch that?

This was (and is) true since the very beginning of the caching objects.
We only create the CachingObject when dispatching the events:
https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apiserver/pkg/storage/cacher/cacher.go#L1026

FWIW - our memory usage would already be 2x+ (or sth like that) higher if it wouldn't be true.

Regarding catching this - we don't have alerting on the memory, but we're doing spot-checking of significant increases, so that would one should be caught.
I agree that we should have more automatic alerting, but can we not block this PR [this is effectively not a potential problems that this PR is introducing - it already existed for the last couple years]

@liggitt - PTAL